extract AST manipulation functions from utils
