+++ packages/server/src/complete/index.ts
@@ -1,5 +1,4 @@
 import { Identifier } from './Identifier'
-import * as utils from './utils'
 import { complete } from './complete'
 
-export { Identifier, utils, complete }
+export { Identifier, complete }
